pyproject.toml
[tool.poetry]
name = "pubmed-paper-fetcher"
version = "0.1.0"
description = "Fetch PubMed papers with non-academic author affiliations"
authors = ["Your Name <pataballapravallika@gmail.com>"]
packages = [{ include = "pubmed_paper_fetcher" }]

[tool.poetry.dependencies]
python = "^3.8"
biopython = "^1.83"
pandas = "^2.0"
pytest = "^7.0" 

[tool.poetry.scripts]
get-papers-list = "pubmed_paper_fetcher.cli:main"
